add testing helpers for api's
